@@ -17,18 +17,18 @@
   "license": "MIT",
   "devDependencies": {
     "@microsoft/load-themed-styles": "^1.2.2",
-    "@types/chai": "^3.4.35",
-    "@types/enzyme": "^2.7.5",
-    "@types/es6-promise": "^0.0.32",
-    "@types/mocha": "^2.2.39",
-    "@types/webpack-env": "^1.13.0",
+    "@types/chai": "3.4.35",
+    "@types/enzyme": "2.8.0",
+    "@types/es6-promise": "0.0.32",
+    "@types/mocha": "2.2.39",
+    "@types/webpack-env": "1.13.0",
     "mocha": "~3.3.0",
     "office-ui-fabric-react": ">=4.10.1 <5.0.0",
     "raw-loader": "^0.5.1",
     "react": "^15.4.2",
     "react-addons-test-utils": "^15.4.2",
     "react-dom": "^15.4.2",
-    "webpack": "^2.4.1",
+    "webpack": "^3.0.0",
     "webpack-node-externals": "^1.5.4"
   },
   "dependencies": {
