@@ -17,14 +17,14 @@
   },
   "devDependencies": {
     "@microsoft/web-library-build": ">=3.0.0-0 <4.0.0-0",
-    "@types/chai": "^3.4.35",
-    "@types/enzyme": "^2.7.5",
-    "@types/es6-promise": "^0.0.32",
-    "@types/mocha": "^2.2.39",
-    "@types/react": "^15.0.25",
-    "@types/react-addons-test-utils": "^0.14.18",
-    "@types/react-dom": "^15.5.0",
-    "@types/webpack-env": "^1.13.0",
+    "@types/chai": "3.4.35",
+    "@types/enzyme": "2.8.0",
+    "@types/es6-promise": "0.0.32",
+    "@types/mocha": "2.2.39",
+    "@types/react": "15.0.25",
+    "@types/react-addons-test-utils": "0.14.18",
+    "@types/react-dom": "15.5.0",
+    "@types/webpack-env": "1.13.0",
     "css-loader": "^0.28.0",
     "enzyme": "^2.7.0",
     "es6-promise": "^4.1.0",
@@ -40,7 +40,7 @@
     "ts-loader": "^2.1.0",
     "typescript": "2.2.2",
     "webpack-dev-server": "^2.4.1",
-    "webpack": "^2.4.1"
+    "webpack": "^3.0.0"
   },
   "peerDependencies": {
     "react": "^0.14 || ^15.0.1-0 || ^16.0.0-0",
