@@ -13,14 +13,14 @@
   "devDependencies": {
     "@microsoft/web-library-build": ">=3.0.0-0 <4.0.0-0",
     "@microsoft/load-themed-styles": "^1.2.2",
-    "@types/chai": "^3.4.35",
-    "@types/enzyme": "^2.7.5",
-    "@types/es6-promise": "^0.0.32",
-    "@types/mocha": "^2.2.39",
-    "@types/react": "^15.0.25",
-    "@types/react-addons-test-utils": "^0.14.18",
-    "@types/react-dom": "^15.5.0",
-    "@types/webpack-env": "^1.13.0",
+    "@types/chai": "3.4.35",
+    "@types/enzyme": "2.8.0",
+    "@types/es6-promise": "0.0.32",
+    "@types/mocha": "2.2.39",
+    "@types/react": "15.0.25",
+    "@types/react-addons-test-utils": "0.14.18",
+    "@types/react-dom": "15.5.0",
+    "@types/webpack-env": "1.13.0",
     "@uifabric/example-app-base": "4.0.0",
     "@uifabric/utilities": "4.4.0",
     "autoprefixer": "^6.7.6",
@@ -39,7 +39,7 @@
     "style-loader": "^0.17.0",
     "ts-loader": "^2.0.1",
     "typescript": "2.2.2",
-    "webpack": "^2.4.1",
+    "webpack": "^3.0.0",
     "webpack-dev-server": "^2.4.1",
     "webpack-notifier": "^1.5.0"
   },
