@@ -3,6 +3,7 @@
   "version": "4.10.2",
   "description": "Reusable React components for building experiences for Office 365.",
   "main": "lib/index.js",
+  "module": "es6/index.js",
   "typings": "lib/index.d.ts",
   "repository": {
     "type": "git",
@@ -10,7 +11,8 @@
   },
   "license": "MIT",
   "scripts": {
-    "build": "gulp",
+    "build-es6": "tsc --outDir es6 --module es6",
+    "build": "gulp && npm run build-es6",
     "clean": "gulp clean",
     "start": "node node_modules/webpack-dev-server/bin/webpack-dev-server.js --config webpack.serve.config.js --open",
     "test": "gulp"
@@ -18,17 +20,17 @@
   "devDependencies": {
     "@microsoft/load-themed-styles": "^1.2.2",
     "@microsoft/web-library-build": ">=3.0.0-0 <4.0.0-0",
-    "@types/chai": "^3.4.35",
-    "@types/enzyme": "^2.7.5",
-    "@types/es6-promise": "^0.0.32",
-    "@types/mocha": "^2.2.39",
-    "@types/prop-types": "^15.5.1",
-    "@types/react": "^15.0.25",
-    "@types/react-addons-test-utils": "^0.14.18",
-    "@types/react-dom": "^15.5.0",
+    "@types/chai": "3.4.35",
+    "@types/enzyme": "2.8.0",
+    "@types/es6-promise": "0.0.32",
+    "@types/mocha": "2.2.39",
+    "@types/prop-types": "15.5.1",
+    "@types/react": "15.0.25",
+    "@types/react-addons-test-utils": "0.14.18",
+    "@types/react-dom": "15.5.0",
     "@types/resemblejs": "~1.3.28",
-    "@types/sinon": "^2.2.2",
-    "@types/webpack-env": "^1.13.0",
+    "@types/sinon": "2.2.2",
+    "@types/webpack-env": "1.13.0",
     "@uifabric/example-app-base": "4.0.0",
     "autoprefixer": "^6.7.6",
     "chai": "^3.5.0",
@@ -63,7 +65,7 @@
     "ts-loader": "^2.0.1",
     "typescript": "2.2.2",
     "vinyl-ftp": "0.6.0",
-    "webpack": "^2.4.1",
+    "webpack": "^3.0.0",
     "webpack-bundle-analyzer": "^2.2.1",
     "webpack-dev-server": "^2.4.1",
     "webpack-notifier": "^1.5.0",
