@@ -13,13 +13,13 @@
     "@microsoft/sp-build-web": "~1.0.0",
     "@microsoft/sp-module-interfaces": "~1.0.0",
     "@microsoft/web-library-build": ">=3.0.0-0 <4.0.0-0",
-    "@types/chai": "^3.4.35",
-    "@types/enzyme": "^2.7.5",
-    "@types/mocha": "^2.2.39",
-    "@types/react": "^15.0.25",
-    "@types/react-addons-test-utils": "^0.14.18",
-    "@types/react-dom": "^15.5.0",
-    "@types/webpack-env": "^1.13.0",
+    "@types/chai": "3.4.35",
+    "@types/enzyme": "2.8.0",
+    "@types/mocha": "2.2.39",
+    "@types/react": "15.0.25",
+    "@types/react-addons-test-utils": "0.14.18",
+    "@types/react-dom": "15.5.0",
+    "@types/webpack-env": "1.13.0",
     "@uifabric/example-app-base": ">=1.3.12 <2.0.0",
     "chai": "^3.5.0",
     "enzyme": "^2.7.0",
@@ -37,7 +37,7 @@
     "ts-loader": "^2.1.0",
     "typescript": "2.2.2",
     "uglifyjs-webpack-plugin": "^0.4.3",
-    "webpack": "^2.4.1",
+    "webpack": "^3.0.0",
     "webpack-bundle-analyzer": "^2.2.1",
     "webpack-dev-server": "^2.4.1",
     "webpack-notifier": "^1.5.0",
