@@ -10,8 +10,8 @@ import {
   BaseComponent,
   KeyCodes
 } from '../../Utilities';
-import * as stylesImport from './Calendar.scss';
-const styles: any = stylesImport;
+import styles from './Calendar.scss';
+// const styles: any = stylesImport;
 
 export interface ICalendarState {
   /** The currently focused date in the calendar, but not necessarily selected */
