@@ -22,8 +22,8 @@ import {
   isInDateRangeArray
 } from '../../utilities/dateMath/DateMath';
 
-import * as stylesImport from './Calendar.scss';
-const styles: any = stylesImport;
+import styles from './Calendar.scss';
+// const styles: any = stylesImport;
 
 const DAYS_IN_WEEK = 7;
 
