@@ -16,10 +16,10 @@ import {
 } from '../../Utilities';
 import { getRelativePositions, IPositionInfo, IPositionProps, getMaxHeight } from '../../utilities/positioning';
 import { Popup } from '../../Popup';
-import * as stylesImport from './Callout.scss';
+import styles from './Callout.scss';
 import { AnimationClassNames } from '../../Styling';
 
-const styles: any = stylesImport;
+// const styles: any = stylesImport;
 
 const BEAK_ORIGIN_POSITION = { top: 0, left: 0 };
 const OFF_SCREEN_STYLE = { opacity: 0 };
