@@ -9,8 +9,8 @@ import {
   ICheckbox,
   ICheckboxProps
 } from './Checkbox.Props';
-import * as stylesImport from './Checkbox.scss';
-const styles: any = stylesImport;
+import styles from './Checkbox.scss';
+// const styles: any = stylesImport;
 
 export interface ICheckboxState {
   /** Is true when the control has focus. */
