@@ -21,8 +21,8 @@ import {
   Icon,
   IIconProps
 } from '../../Icon';
-import * as stylesImport from './ContextualMenu.scss';
-const styles: any = stylesImport;
+import styles from './ContextualMenu.scss';
+// const styles: any = stylesImport;
 export interface IContextualMenuState {
   expandedMenuItemKey?: string;
   dismissedMenuItemKey?: string;
