@@ -22,10 +22,10 @@ import {
   IDragDropOptions,
 } from './../../utilities/dragdrop/interfaces';
 import { IViewport } from '../../utilities/decorators/withViewport';
-import * as stylesImport from './DetailsRow.scss';
-const styles: any = stylesImport;
+import styles from './DetailsRow.scss';
+// const styles: any = stylesImport;
 import { AnimationClassNames } from '../../Styling';
-import * as checkStyles from './DetailsRowCheck.scss';
+import checkStyles from './DetailsRowCheck.scss';
 
 export interface IDetailsRowProps extends React.Props<DetailsRow> {
   item: any;
