@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { IColumn } from './DetailsList.Props';
 import { BaseComponent, css } from '../../Utilities';
-import * as stylesImport from './DetailsRow.scss';
-const styles: any = stylesImport;
+import styles from './DetailsRow.scss';
+// const styles: any = stylesImport;
 
 const INNER_PADDING = 16; // Account for padding around the cell.
 
