@@ -3,8 +3,8 @@ import { BaseComponent, css } from '../../Utilities';
 import { IDocumentCardActionsProps } from './DocumentCard.Props';
 import { Icon } from '../../Icon';
 import { IconButton } from '../../Button';
-import * as stylesImport from './DocumentCard.scss';
-const styles: any = stylesImport;
+import styles from './DocumentCard.scss';
+// const styles: any = stylesImport;
 
 export class DocumentCardActions extends BaseComponent<IDocumentCardActionsProps, any> {
   public render() {
