@@ -11,8 +11,8 @@ import { Icon } from '../../Icon';
 import { GroupSpacer } from './GroupSpacer';
 import { Spinner } from '../../Spinner';
 import { FocusZone, FocusZoneDirection } from '../../FocusZone';
-import * as stylesImport from './GroupHeader.scss';
-const styles: any = stylesImport;
+import styles from './GroupHeader.scss';
+// const styles: any = stylesImport;
 
 export interface IGroupHeaderState {
   isCollapsed: boolean;
