@@ -8,8 +8,8 @@ import {
   getNativeProps
 } from '../../Utilities';
 import { ILink, ILinkProps } from './Link.Props';
-import * as stylesImport from './Link.scss';
-const styles: any = stylesImport;
+import styles from './Link.scss';
+// const styles: any = stylesImport;
 
 interface IMyScreen extends Screen {
   left: number;
