@@ -8,8 +8,8 @@ import {
   OverflowSet
 } from 'office-ui-fabric-react/lib/OverflowSet';
 
-import * as stylesImport from './OverflowSet.Example.scss';
-const styles: any = stylesImport;
+import styles from './OverflowSet.Example.scss';
+// const styles: any = stylesImport;
 
 export class OverflowSetCustomExample extends BaseComponent<any, any> {
 
