@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { BaseComponent, css, DelayedRender } from '../../Utilities';
 import { ISpinnerProps, SpinnerType, SpinnerSize } from './Spinner.Props';
-import * as stylesImport from './Spinner.scss';
-const styles: any = stylesImport;
+import styles from './Spinner.scss';
+// const styles: any = stylesImport;
 
 export class Spinner extends BaseComponent<ISpinnerProps, any> {
   public static defaultProps: ISpinnerProps = {
