@@ -15,8 +15,8 @@ import { ITooltipHostProps, TooltipOverflowMode } from './TooltipHost.Props';
 import { Tooltip } from './Tooltip';
 import { TooltipDelay } from './Tooltip.Props';
 
-import * as stylesImport from './TooltipHost.scss';
-const styles: any = stylesImport;
+import styles from './TooltipHost.scss';
+// const styles: any = stylesImport;
 
 export interface ITooltipHostState {
   isTooltipVisible?: boolean;
